Transfer PipContentOverlay when swipe to home with Shell transition
Swipe to home animation is played by launcher, but the entering PIP
transition will finished by PipTransition. As a result, we need to
reparent the overlay and fade out the overlay in PipTransition when
enter PIP is finalized, otherwise the overlay will be removed with the
remote transition leash and cause flicker.

Bug: 222030101
Test: verify with swipe to home with Shell transition
Change-Id: Idde3414e3cc5e8193cf7d4da9818e59dc6b4b1c5

Former-commit-id: b6a5bf31ce90496bec330e97a782b674ff3e2b33