Remove AndroidKeystore key migration code in framework
Since there will be no appId transitions in T, AndroidKeystore keys do
not need to be migrated. Also, even in the future when appId
transitioning is allowed, key migration will not be needed as it is
possible to select a sharedUserInheritor to have its appId persisted,
which means keys will also be inherited implicitly.

The conclusion is that the functionality of migrating keys across the
UID namespace is no longer needed, even in the future.

Test: m
Bug: 220015249
Change-Id: I6d0976fea095ec67292642d20ef8014172e83b2e

Former-commit-id: 1b81d4d28479b9572b83f16fbd83dd4d682f31b5