Add error log for adding parented complications.
Adding a complication view that already has a parent throws an
IllegalStateException. This change adds an error log for that case that
includes the class of the complication to make it easier to track down.
The most likely cause of this is complication view reuse. Complication
views should be created anew each time they are needed.

Test: manually
Bug: 220478774
Change-Id: Icaa070ae3c852e89e7203e4b3ff3ee42bf515ce0

Former-commit-id: 5a2db6da7a5e659db2c87cc42fc39ecaef85b5e0