SystemUI: Fix signal bar icon overlay issue
If the signal bar visible state is STATE_DOT, it should be shown
a dot instead of signal bar icons. But the signal bar icon will
be shown again if the mobile state changed, and then, the dot and
signal bar icon are both shown. So don't update signal bar
visibility if the state is not STATE_ICON

Bugs: 200915946
Test: Manual
Change-Id: I7fb04049790f112441511462d3f2963aed42c790

Former-commit-id: d39037f684cdc0051bfa6fb05332320c36cff583