Cancel touches on the unlocked animation
On the lock screen, users are able to drag
their finger to the UDFPS location which can leave
the panel expansion in a transition state while
the device is being transioned to the home screen.
therefore, we need to:

* on unlock, instantCollapse notification panel if it's already in the
middle of being flung away so there aren't overriding animators.

* don't expand the bouncer if the device is unlocking

* cancel launch transition end runnables on going to sleep
in case the user presses the power button quickly after
unlocking, before the launch transition end runnable gets
to run. We don't end up in a bad state where we hide the
keyguard (due to the launchTransitionEndRunnable) when we
actually wanted to show it

* don't update the UNLOCKED scrim when the device is
updating the panel expansion during the unlock animation

Fixes: 222324260
Test: atest SystemUITests
Test: manually authenticate UDFPS from dragging finger
from diagonally above/right of UDFPS (and from below)

Change-Id: I8f0b50c01544c3cf206bb764b2c5e5554dcd256d

Former-commit-id: df83af614fa159a7df519afcf8fac757d3852b07