Transfer launch cookie when finishing activity
For example: launcher starts a task T which contains activity A
with launch cookie. And after A was launched or used a while, it
launches B and finishes itself. When closing task T, it still
expects that launcher can receive the same launch cookie of the
task, then launcher can animate such as returning to the bounds of
widget. So this change allows B to transfer its launch cookie to
A when B is going to finish.

Fixes: 220290671
Test: atest ActivityRecordTests#testTransferLaunchCookieWhenFinishing
Change-Id: I520dc696de64c4e750be607cc03c843499508269

Former-commit-id: 3ad8e00f86e777f8c498d61e261554e4b54e9cae