Ensure blur radius goes to 0
If the animation is canceled, it may not end the blur radius at 0,
which affects NotificationShadeWindowControllerImpl for wake and
unlock scenarios. When this is not 0, SysUI then believes it should
still be visible and retain focus, which absorbs all touches that
should be going to launcher. It was also noticed that some users has
reported a blurry background.

Bug: 220703613
Test: manual (wake and unlock over and over again)
Change-Id: Ief7ba5eabbcbbe08e2f2dec3072420aa5539fa8f

Former-commit-id: 2ad14d184e622057925605d38ec4542468717698