Offset animation layer for TaskFragment open/close
It is possible to have a fullscreen closing TF between two resume TFs
(this can happen when closing the fullscreen TF results the bottom
activity open a new activity into split). In that case, we don't want
the fullscreen TF to hide the activity below it during cross fade
animation.

Fix: 214175523
Test: manually verify with Settings
Change-Id: I4350114d800442308c9c53f474ba60542c8a86e5

Former-commit-id: c562f322b21067b251cb3336878c471036126ae4