Allow profiling, dumpheap, attach-agent for sdk sandbox
This CL adds cases for the sdk sandbox when performing debuggable checks
for profiling, dumpheap and attach-agent adb commands. To achieve this,
the CL adds a getClientInfoForSdkSandbox method to the ProcessRecord.
isDebuggable() has also been added as a method to ProcessRecord to make
the checks readable.

To support profiling, the CL also ensures the correct zygote flags are
set when the client app of a sdk sandbox is profileable/debuggable.

Test: w/ debuggable client app (adb shell am dumpheap sdk_sandbox_[x])
Test: w/ debuggable client app (adb shell am attach-agent
sdk_sandbox_[x] [filepath])
Test: w/ debuggable/profileable client app (adb shell am profile start
sdk_sandbox_[x] [filepath])
Fixes: 222670584
Fixes: 221004701

Change-Id: I72bf66906ba1b8283474345562ed43fdf433e793

Former-commit-id: ca1205969908e07ce0c78a4a51177314f89fbfe4