Fixs the window bounds not correct after rotation
When the device is rotated and the window is magnifying, we need to
apply the displayInfo matrix from the SF's callback to obtain the
window bounds based on the rotated-coordination and the inverse matrix
from the magnification spec to obtain the un-magnified window bounds.

The order of applying the displayInfo matrix and inverse matrix should
be the former first and then the latter, and we can get the correct
window bounds.

Bug: 223687400
Test: a11y CTS & unit tests
Test: Manual testing including the A11y services
Change-Id: I72d14170b1ecd88c8b28ba7433dc8bee8626dfa0

Former-commit-id: b63928a256f5d578979479efb6727dcb87a86d87