Check instrumentation source UID in setInstrumenting
In order to allow switching touch mode, WindowManagerService checks if caller
has MODIFY_TOUCH_MODE_STATE permission OR if the process is instrumeting.

Just checking if process is instrumenting may not be enough since apps
can self instrument themselves. To avoid that, we're now also checking if
the instrumentation source is shell (which has the permission pointed
out above granted) or root.

Bug: 222652344
Test: atest ActivityTaskManagerServiceTests
Change-Id: I55078d5c223d4f936d58199d3d013e07a6747a1e

Former-commit-id: a5e0a4f770ca1ca0163be4b367e2ae1747ba1758