Fix finishTranslation doesn't work if start Activity in the same task
If the app starts a new Activity in the same task then the
finishTranslation() is called, the finish translation doesn't work
because our logic is to check if task top Activity is our expected
translation Activity. In this case, the top Activity is the new
Activity, the original Activity is paused in the same task. To fix
the problem, we get all non-finishing Activities in the task and
check if the target Activity in this task.

Bug: 202570916
Test: atest UiTranslationManagerTest
Test: atest CtsVoiceInteractionTestCases
Test: atest CtsAssistTestCases

Change-Id: I2471e04b894890777e13dd45c290f059dbe5f487

Former-commit-id: b9fe1923d9622d5deaef1ed62685d1ece797ae7a