Remove unnecessary performLayout
We don't do layout if nothing is changed. The insets system has changed
significantly. The original issue might not exist. This CL removes the
unnecessary performLayout.

This is also a step to move window layout to the client side. When that
happens, we won't have performLayout on the server side.

Bug: 112205777
Bug: 161810301
Test: Rotate phone, swipe up and ensure that home gets the right insets
Change-Id: I4289805e5aa10e82412c7233b323fbb9a7d84e33

Former-commit-id: d2c8fc05366f50057d5de71e8509f4bb20fdf78f