Add support for TEAMFOOD flag.
A list of boolean teamfood flags is now available. Adding a flag to
the list means that the flag is tied to the value of the teamfood
flag.

If the flag is manually set, that manually set value will take
precedence.

If the flag is true by default, the true value will take precedence.

Otherwise, the flag will fallback to the value of the teamfood flag.

Bug: 223379190
Test: atest SystemUITests
Change-Id: Id4739679e633c25abccd9093d15eb5e016cd5ac2

Former-commit-id: 214c0014ef4f12b93d9ef33f0a94f1312cce7f50