Only note restrictions attempts for RESTRICTED bucket.
We don't want to put an app in the RESTRICTED bucket too soon after
a user has used the app. If the system attempts to restrict the app
(probably through the restrictApp() API), and the user has recently
used the app, then we'll note the restriction attempt and delay
putting the app into the bucket until our min delay time has passed.
Now, we make sure to only note the restriction attempt when the bucket
is the RESTRICTED bucket.

Bug: 223863858
Test: atest FrameworksMockingServicesTests:AppStandbyControllerTests
Change-Id: Iaf6647bfd8a72206f2b80df8b21d2446f4eeb291

Former-commit-id: 250f9f9099396bac3c52b4d8030ef949debfd7fa