Fix an issue where bubbles didn't collapse onTaskMovedToFront
I noticed when opening settings while stack is expanded wouldn't
cause the stack to collapse. This was because we skipped the collapse
if the shade was expanded. I don't think there's any reason to be
doing this, I think this was accidentally added during the refactor
to move bubbles into shell.

Test: manual - have bubbles expanded, pull down the shade, tap on
               the settings gear or long press on a tile, ensure that
               when the settings screen opens that the stack collapsed
Bug: 223635739
Change-Id: Ia3476e694a8e64c95eed63dea4a3729e83766683

Former-commit-id: ba395508c583a852d91f0f5895ccb75a8d8d0502