Use ClientWindowFrames to collect frames computed by WindowLayout
This is a step to move the window layout to the client side.

ClientWindowFrames is used to carry window frames dispatched from the
server to the client for now. Later, the window frames will be computed
at the client side, and the frames will be sent to the server side via
ClientWindowFrames.

Bug: 161810301
Test: presubmit (no behavior change but only code refactors)
Change-Id: I83d8cfff2432e09759ef5b3d5f3b21731fc71bff

Former-commit-id: 45790587c729db89fc366d85ba45bc7081ea7d9c