Fix last notification trimmed on lockscreen
The stack height set to ambientState, and then used by `StackScrollAlgorithm#initAlgorithmState` had some additional padding. For this reason, in `initAlgorithmState()` the `state.firstViewInShelf` was actually the one next to the correct one.

intrinsicContentHeight doesn't have those paddings, and result in using the correct size. This can also be double checked by enabling debug draw lines in NotificationStackScrollLayout.

The test is deleted as it was just duplicating the method code.

Bug: 214504318
Test: existing tests pass + manually verified
Change-Id: I83179d56ca7299e5d020e4eea2d460b168a76c6e

Former-commit-id: c060228cc2726a9fc99e23b986ad324ef4f9b758