Generate an abort event when stopping
Because the client callbacks from SoundTriggerMiddleware service are
async, if a client quickly stops and starts a model, then receives a
detection event for that model, it is impossible for it to tell
whether the event corresponding to the previous or current session,
and thus is unable to reason about the resulting state of the model.

To mitigate that, we will always send an abort event if a model is
stopped while it was in an active state. This way the client can
consider the model active (even after stop()) until an event is
received for that model.

Bug: 191935600
Test: Manual verification of soundtrigger use-cases.
Test: atest SoundTriggerMiddlewareImplTest
Change-Id: I59555b024df54ca90ae7573acb38e114dac0c130

Former-commit-id: 3b660525f37bdb6c549d9ce483658d4a8f189438