Fix the flaky AccessibilityMagnificationTest#testListener
The listner might invoked twice because we use vsync to
change the magnification spec. When the animation is running,
the posted tasks might executed after animation ended twice.

To fix, we update the source bounds only when it is indeed changed,
which also could reduce the redudant mirroing content operation when
the request is sent consecutively in a short time.

Bug: 214785717
Test: atest AccesibilityMagnificationTest \
      --retry-strategy ITERATIONS --max-testcase-run-count 20
Change-Id: I68962b189d8f39f587b717e8125a224e0f67a7e7

Former-commit-id: d7219b7d60b6905c1941bc55f92b0228a7955aab