Preserve task when clearing top activities
There is no guarantee that clear-top won't finish all activities
in the task, such as the last activity is declared as default
launch mode and the intent flags don't have SINGLE_TOP, then it
will also be finished.

Otherwise when adding the new activity into the detached task,
it will cause NPE in onChildPositionChanged because its display
content is null.

Fixes: 224462770
Test: TaskTests#testPerformClearTop
Change-Id: Icfa0390a5881043bc55486347aa2854db100d171

Former-commit-id: 56ce0fa5c1777ef0a3aec6d9b3495d4bdfe3b29d