Check whether a display supported pip when an activity try to enter pip
ActivityRecord#supportsPictureInPicture() will also be called when an
activity sets PictureInPictureParams, this cause an issue if an activity
is trying to set pip params but it doesn't enter pip.

We should only block an activity enter pip on specified display instead
of telling caller that the Activity doesn't support pip.

Bug: 224699616
Test: manually, atest WmTests:ActivityRecordTests
Change-Id: Id8eb421ebab57e6c022650f3e820175531a2e8fc

Former-commit-id: 2c2516afc79d34a1b91c5c250faea69ef712f52e