Add NPE check in ColorFade.
There is a delay between when a display is removed in the
DisplayManager thread and when the ColorFade is dismissed in the
PowerManager thread which can leave ColorFade in a state where its
associated display is no longer available. Check for valid display
similar to how we are doing in other parts of the ColorFade
code.

Bug: 220650894
Test: atest ColorFadeTest
Change-Id: I75528bf4cd398999e012c3c8d8c53b6ac0ce1552

Former-commit-id: ba75b774fa45487f820e3a55765ecb9b820d18fc