Wait for an abort event when stopping a model
Because the client callbacks from SoundTriggerMiddleware service are
async, if a client quickly stops and starts a model, then receives a
detection event for that model, it is impossible for it to tell
whether the event corresponding to the previous or current session,
and thus is unable to reason about the resulting state of the model.

To mitigate that, we will always wait for an abort event when a model
is stopped before any other operation is attempted. This way the
model state stays synchronized.

Fixes: 191935600
Test: Manual verification of soundtrigger use-cases. Specifically,
      quick toggling of Assitant and Now Playing via Settings.


Change-Id: I010558e8fa1891922f20dcdb7b0ca04c9500c2e1

Former-commit-id: 6e2eb81ac4687ac5340673ff996c0591154def7f