Don't expand shade when pulsing w/o tracking notif
When pulsing, we only allow dragging down anywhere to expand
the shade when the keyguard can be bypassed. Else, the user
needs to drag the notification itself to expansion.

Why? Devices with UDFPS can wakeup the display to the pulsing
state without notifications. In this state, users shouldn't be
able to swipe to expand the notification shade while pulsing.

Test: atest SystemUITests
Fixes: 217247200
Change-Id: Ieb6bf058bac95a4e7259703d1e87e99f910b40c8

Former-commit-id: 97834b3c303f31ed5abb5f89566e50e8911a467f