DeviceConfig for Chooser update-delay duration.
For more information, see go/sharesheet-list-view-update-delay.

Bug: 224816815
Test: `atest ChooserActivityTest`. The duration can be manually
configured by `adb shell device_config put systemui
sharesheet_list_view_update_delay n` to set a delay of n milliseconds.
With a timeout of 250 (as it was before this change), all the tests
still pass. With a timeout of 0, all tests pass but one (as described in
the doc linked above) -- a potential fix for that test will follow in a
subsequent CL.

Change-Id: Ia5edfe391b135659d2fd8126069c5cd6ce15653b
Merged-In: Ia5edfe391b135659d2fd8126069c5cd6ce15653b

Former-commit-id: 944c5af2d685a8fc31e4d79c7207ba6bedbb58db