Consolidate finally block of startActivityInner
The concept is similar to change
  if (c) fun1(); else fun3();
  if (c) fun2(); else fun4();
to
  if (c) {fun1(); fun2();} else {fun3(); fun4();}
So the logic for the same condition can be put together.

Besides, finally block should not be too large because compiler
just duplicates the entire block.

It is also better to put any operations that might change the
hierarchy before calling continueWindowLayout.

Also fix format and missing info of dumping ActivityStarter.

Bug: 163976519
Test: CtsActivityManagerDeviceTestCases

Change-Id: I8f6ca71fc9a13296fa391aef6656d533908adffb

Former-commit-id: 14185cd0eed7ca9043c5aa8115e90594db6b4069