Use composition size limit on repeating effects
Concatenate short repeating vibration patterns for PWLE and composed
primitives into a larger command to the vibrator HAL, making use of the
limits for compositions.

Add a check to find the best place to break down a large or repeating
PWLE waveform, preferably at amplitude zero, to avoid segmenting the HAL
commands on a high amplitude, causing a click effect.

Fix: 224930189
Test: VibrationThreadTest
Change-Id: Iac59855886596c98de583690356ef02bc67791df

Former-commit-id: b2bb2eda0fd1ca3057049f6048402d8a38016ffe