Change Ethernet API to use OutcomeReceiver
This change addresses API review feedback. In addition, it changes the
returned result from Network to the interface name. The current API
returning a Network object is racy and cannot be implemented correctly.
Users should instead use the ConnectivityManager#requestNetwork() API to
get hold of the Network for a given interface.

CP note: the Merged-In ID is not the actual Change-Id of this change in
tm-dev. This change was originally merged in aosp, but was subsequently
reverted and cherry-picked into tm-dev. The cherry-pick falsely reused
the Change-Id of the (now reverted) original change. This causes
automerger to skip this change on all branches between aosp-master and
tm-dev when the correct Merged-In ID is used. Therefore, the Merged-In
ID listed is the Change-Id of an unrelated change that is present in
tm-dev but not aosp-master.

Bug: 220017952
Test: TH
Ignore-AOSP-First: CP to make automerger happy.
Merged-In: Iefd24d955572589c5bd3ca9a8139ea6e44979e8a
Change-Id: I7c46545a47034be409071c2ec007d9e1480c6ed0

Former-commit-id: 954e0d86cfcd7866b1b6dc3c38b4ddf61633a434