Change Ethernet API to use OutcomeReceiver
This change addresses API review feedback. In addition, it changes the
returned result from Network to the interface name. The current API
returning a Network object is racy and cannot be implemented correctly.
Users should instead use the ConnectivityManager#requestNetwork() API to
get hold of the Network for a given interface.

Bug: 220017952
Test: TH
Ignore-AOSP-First: CP to make automerger happy.
Change-Id: I7c46545a47034be409071c2ec007d9e1480c6ed0

Former-commit-id: b96b2c755fdae6967f4d070741304477403005e3