Overlay config: include config file data in dumpsys
Overlay packages are configured via one of two mutually exclusive
approaches: either via the overlays' manifests or via XML config files.
The overlay manager dumpsys already includes data for the manifests;
update the output to also include config file data.

The new output includes the raw XML, but to save memory, this is omitted
on user builds.

Example output; note the parsedConfigFile field, and that the parsedInfo
field is null.

  0, ParsedConfiguration{packageName=com.android.overlaytest.framework, enabled=true, mutable=true, policy=vendor, parsedInfo=null, parsedConfigFile=ParsedConfigFile{path=/vendor/overlay/config/config.xml, line=2, xml=<overlay package="com.android.overlaytest.framework" enabled="true" />}}

Bug: 224945112
Test: atest OverlayConfigTest
Test: adb exec-out cmd overlay dump
Change-Id: I1d880d2cc6cff763865b9b445fd2caa7d35c431f

Former-commit-id: fa61d140dab1d08074d7b0f6440aa5341ed5e041