Fix pip finish-callback when enter animation runs in shell
PiP animations happen in both launcher and in shell. On top of that,
the implementations are different!? The launcher impl will actually
merge expand-pip into an enter animation, but the shell impl waits
for the enter animation to finish. This means we need to alter the
finish-handling logic depending on whether the animation is running
"locally" in shell or not.

Additionally, fixed an ordering-problem: This was calling the
finish-callback before unsetting the mFinishCallback member. The
problem is that if there are pending animations queued-up, those
will start in sync -- if those end up calling-back into PipTransition,
then they will see mFinishCallback as non-null.

Bug: 224622288
Test: atest PinnedStackTests
Change-Id: Ie3696b101ecca69fae93b347ff358863cdb4e6c3

Former-commit-id: 9c464cc063c78e164db0d84c2b7c189d97e3dd88