Add extra to inform the role holder of the update attempt result
Role holders can request an update. When that happens, the flow
returns back to platform, platform performs the update and checks
the update result. If update was successful, it restared the role
holder, and if not, it failed provisioning. This CL updates this
so that the role holder is _always_ started, regardless of the
update result. That way the role holder can make a decision
whether to continue provisioning or fail it.

Bug: 224993192
Test: none
Change-Id: Ia7dd61d02c568e8c8ad17af0a61106d127c4869a

Former-commit-id: 4fe4b2e65d68f09bf14601b0ab3b3ea90cc1ed4c