Return START_TASK_TO_FRONT for bring fullscreen task
The condition was added for split screen that intends to return
START_DELIVERED_TO_TOP for an activity on top of the split but
not the global top. While in fullscreen mode, it is possible
that a translucent task on top of another one. That should still
return START_TASK_TO_FRONT when moving the bottom visible activity
to top.

It may affect shell transition because only START_SUCCESS or
START_TASK_TO_FRONT requires transition, otherwise it will abort.

Bug: 223988300
Test: atest ActivityStarterTests#testMoveVisibleTaskToFront
Test: Enable shell transition.
      Launch a landscape translucent activity from launcher.
      Launch home again and no crash by
      "Trying to rotate outside a transition".

Change-Id: I2073084d5e7318b114fa74718caf4315fb21fa83

Former-commit-id: d23fb21ab3888c999630b9572fc46707ba85faa5