Restore long press power behaviour setting
The LPP setting is a bit special - the behaviour is configured by the config.xml and can then be modified via user setting to invoke Assistant if available. When restoring this value we need to be careful to not clobber default config.xml device settings if the user can't change it later. We strongly prefer to rely on config.xml device value in any case.

Bug: 215540406
Test: atest SettingsProvider
Change-Id: I8898a52a5d28f0957f44f3075498b1139c8ce831

Former-commit-id: 1a85be89c74ecbbea4dc7d5bb84751b775edee99