Keep wallpaper on top when finish a non-fullscreen unoccluding window
Compensate isKeyguardUnoccluding does not work for shell transition
because there is no app transition. There should check this
condition by isTransition.

Also fix a generic issue that the wallpaper can become invisible for
a short of time when showWhenLocked window take effect. The
notification shade is the wallpaper target in the beginning, but
it will become invisible because the opening activity will occluding
it, so wallpaper will be hide when notification shade goes to relayout
as invisible.
If there has set useTopWallpaperAsTarget in previous traversal, keep
wallpaper visible when wallpaper target is going away.

Bug: 209906849
Test: for legacy: verify no flickering when finish a showWhenLocked
activity on lock screen.
Test: for shell-transition: verify wallpaper stay visible when finish
a showWhenLocked activity on lock screen.
Test: run "atest KeyguardTests" for both legacy and shell-transition

Change-Id: Iba4cc27fa323b3ddb7ebdaa9d89a1548c4bd1906

Former-commit-id: 741327997e8b7807aa8e3520719fe0102b74b206