User Switcher: Prevent flickering for user switch.
There is a scrim flickering when opening activity from keyguard and from
QS in keyguard.

From Keguard, there is a flickering of the back scrim.
From QS, there is a flickering of the back scrim and the notification
scrim.

In order to prevent this, set alpha of these scrims to 0 when state is
KEYGUARD or SHADE_LOCKED

Bug: 223753460
Test: Manual on relevant devices
Change-Id: Iadf199bb5905de4d9c33e6f01a3fb20993559d25

Former-commit-id: f8f6deb165d1372c521f131fb4b07a35811be009