Register KeyguardNotifVisProvider CoreStartable
KeyguardNotificationVisibilityProvider's `start` method was never being
invoked, because it was never being registered as a CoreStartable. This
was in turn causing all invalidaton events to be missed, resulting in
stale notification shade states being visible.

This CL properly registers the CoreStartable, and adds additional tests
to verify that listeners are called as expected once registered.

Fixes: 224763592
Test: atest KeyguardNotificationVisibilityProviderTest
Test: 0. Have silent notifications
      1. lock device
      2. expand shade, keeping device locked
      3. collapse shade
      4. verify silent header no longer visible
Change-Id: I70661583f0d98c6dac7924667dad89a3664334e2

Former-commit-id: 48b5a61f6de448114a85dfbda6cc95e7f9e756be