Sync PiP WM bounds to rotation immediately for fade-in
When pip is fading-in, it should be reconfigured at the start
of the animation. This means that if the display rotates, we
want to immediately update the pip bounds to match otherwise
the window content will be relayout-ed in the wrong place on
screen -- usually resulting in unwanted insets.

In this case, keep track of when we are expecting to start a
fade-in and, in that situation, recalculate the bounds.

Additionally, fix a non-deterministic ordering situation where
some legacy per-activity callbacks would make global changes.
This meant that the order of activities in mParticipants could
effect which transition was actually requested. This CL moves
all the legacy callbacks after the other per-activity logic
in finishTransition.

Bug: 222033492
Test: Open landscape auto-pip app. Open recents, tap on edge
      watch pip fade-in without content-flicker
Change-Id: Iea3dc3ac4ddda02e1b464baedd11fb1cc9cbe930

Former-commit-id: 433c5eae073416d7b30536b0335f91018443985c