Re-deprecate animation backgroundColor API
It's prefered to keep the API deprecated since the API was previously
deprecated and made a noop. Undeprecating it would lead to inconsistent
behavior across android versions. With the API not working on some
versions which can easily look like a bug from a developer's
perspective.

Instead we create a new API for the animation background effect refering
to an animation backdrop instead of a background. This also required
updating other API references of background in this context to backdrop.

Bug: 226122431
Test: atest CtsWindowManagerDeviceTestCases:ActivityTransitionTests

Merged-In: I4493d22a33606d639d498b6dc0f6f1bfa3d7f135
Change-Id: I4493d22a33606d639d498b6dc0f6f1bfa3d7f135

Former-commit-id: 2b770a7e86f7d5f7c10f7b77a136bd15592be08a