Fix PIN lock pattern not respecting user settings
The LockPatternView had a boolean attribute mEnableHapticFeedback that
held a copy of the user settings for enabling "Touch feedback"
vibrations. This attribute was used together with
HapticFeedbackConstants.FLAG_IGNORE_GLOBAL_SETTING deprecated setting to
bypass the user settings in the vibrator service.

This change removes the use of this deprecated flag and lets the
Vibrator service apply the user settings for "Touch feedback" to
vibrations generated for the PIN lock input in LockPatternView.

Fix: 216656312
Test: manual
Change-Id: If9733c059e5059b6e58b6f84d100df2a0abafab6

Former-commit-id: 9bab88b20ac716630da52808337886f33a82bb08