Remove unnecessary (and potentially problematic) call of notifyDefaultDisplayCallbacks.
Since the KEYGUARD_TIMEOUT message is sent to the handler before the
SHOW message, when handling KEYGUARD_TIMEOUT, mShowing is still false.
The notifyDefaultDisplayCallbacks is later called again in
setKeyguardShowing, after showing has been updated to true.

Also updated setKeyguardShowing to not call
notifyDefaultDisplayCallbacks when only aodShowing is different.

Bug: 221866693
Fix: 221866693
Test: manually on device with instructions in the bug.
Test: manually on device with added logging.

Change-Id: Ie43b1c629e94307a617fcf57c9ebc4ff3e982fbf

Former-commit-id: ac8355b60e3e4edbbb08a01f9174e2e95a91e699