Validate ancestor class for @EnforcePermission
The lint was previously ensuring that any descendant class (or method)
would have the same annotation as the parent. However, it did not verify
that each parent was also annotated. This can occur if a developer
annotates the Java implementation but not the AIDL definition. Add a
verification that if a class or method is annotated, so is its ancestor.

Test: atest --host com.google.android.lint.EnforcePermissionDetectorTest
Bug: 220214993
Change-Id: Ie67fb2f47460da8600865b5ac24bb293dbd4cc6e

Former-commit-id: 0f714a306317ee0e8882e9c95e7124056b954415