Fix notification group scroll in split shade
QsContainer was matching the entire screen size for split shade, while only the qs area in normal shade.

This made `isInsideQsContainer` to always return true for the split shade (also because only the Y coordinate was checked, as there was the assumption QS where only at the top).
Now, the `isInsideQsContainer` (renamed to `isInsideQsHeader`) checks if the touch is inside the QsHeader, that matches the wanted bound both in the normal and split QS.

Fixes: 225328865
Fixes: 223951547
Test: atest NotificationStackScrollLayoutTest
Change-Id: Ie468610eae68eb698310a95a5c9f9d98be4d89fb

Former-commit-id: 48c0ae962c73dfee1dff398701844bdb46d21c5d