Add new TelecomManager method to check if a self-mgd app is in a call.
Adding a new method which NotificationManager will use to check if a
VOIP app is currently engaged in any self-managed calls.  Structured to
accept a package name and userhandle rather than a PhoneAccountHandle so
that NotificationManager doesn't need to query all phone account handles
to find the right one to pass in.

NotificationMgr is part of the system service, so this @hide method is
sufficient.  We can consider making this a system API in the future
when we mainline Telecom.

Test: Manual test with Telecom test app to validate method behavior.
Test: Added new unit test to exercise backing functionality for new method.
Bug: 226959797
Change-Id: I74c0568ad2ec308fe50cada7315e3a677363cccc

Former-commit-id: 27bf43460a324af01c086c9b2afcc2118dd849f2