Separate out bouncer updates
Some callbacks want to know when the bouncer
starts transitioning to show (ie: UDFPS animations)
while other callbacks only care if the bouncer
actually fully showed (ie: running face auth).

We want to make sure we don't over-trigger
face auth, so only request face auth when
we know the bouncer is fully shown rather
than in the middle of a transition to potentially
showing the bouncer (sometimes it can be quickly
cancelled).

Test: atest SystemUITests
Fixes: 226967764
Change-Id: Ifca6fb1861992a78815066571019226322260279

Former-commit-id: c89796fd60f233daf23182c5039d004f55cbc2d1