Restore V2+ signature requirement for system apps
In Android R, the platform began enforcing a requirement that apps
targeting this release or later must be signed with at least the V2
signature scheme. Initially, this requirement was applied to system
apps, but due to the build system stripping V2+ signatures due to
compressed dex and library files, the requirement had to be relaxed
for system apps; since then, the system apps have been updated with
uncompressed dex and library files to prevent the signature stripping
by the build system. This commit restores the requirement that
system apps targeting SDK version 30+ must be signed with at least
the V2 signature scheme.

Bug: 215046612
Test: Manually verified fix properly identified system app without
       V2+ signature
Change-Id: I5b482931a99ea8c2308d516912317ff0828153e2

Former-commit-id: 7cb3a32293ef72498a1b44b11a64341c11d941b8