Send up-to-date InsetsState to the client after rotation
When LOCAL_LAYOUT is enabled, clients compute the window frames on their
own. However, for windows which provide insets, the server computes the
window frames for updating the InsetsState. In this way, all the clients
can receive the up-to-date InsetsState after rotation before the clients
of insets source windows report their frames.

Before LOCAL_LAYOUT is enabled, the server would layout all the windows
after rotation. So the clients receive the up-to-date InsetsStates as
well. The logic remains the same in this CL.

Bug: 161810301
Bug: 175861127
Test: Perform fixed rotation, seamless rotation, and regular rotation.
      See if there is anything unexpected.
Change-Id: I23246f363d5d9ef974a4f7a64041bc9d56042ade

Former-commit-id: 552e7363915e9fd1e45e159c24198277826b1b72