Disable volume-key-accessibility shortcut upon exiting setup wizard
The volume-key-accessibility shortcut is often triggered accidentally,
which causes much frustration for users who are not familiar with it,
and do not know how to revert.

To help prevent this from happening, we will set the shortcut targets
to empty if setup has been completed without the enabled
AccessibilityService associated with the default shortcut.

Bug: 226866175
Test: manual test:
     1. activate shortcut during setup wizard then finish setup wizard,
	observe settings value of shortcut targets. It should not be
	empty.
     2. Finish setup wizard then observe esttings value of shortcut
	targets. It should be empty.
     3. Create a new user and run step 2 and 3 separately.
Change-Id: Iaf99b850043f7606d6b183f0b4822c36287fa106

Former-commit-id: 7efe0b9bd1515d1618fa6728b467f4dfa9b82712