Add perm split W_E_S to A_M_L and sort
Add a permission split:

    WRITE_EXTERNAL_STORAGE to ACCESS_MEDIA_LOCATION

...and sort split permissions.

The reason adding this is that the split-permission handling code
doesn't handle recursive split-permission mappings, but only direct
mappings.

The reason for sorting is that, given a tree of permission splits,
the entries in platform.xml must be sorted topologically, due to
current permission split code.

Bug: 227240947
Test: atest CtsPermissionTestCases CtsPermission3TestCases SplitPermissionsSystemTest
Change-Id: I62c2ab8687af2e8325fcc370a74af0f589d8e9f2

Former-commit-id: 5b9f6a592c5898f7b53b8550780caa2a2002922c