Add a DefaultMixedHandler handler combined transitions
The current TransitionHandlers have generally isolated themselves
based on feature. Transitions, however, are system-wide and thus
several state-transitions involve a mix of windowing features.

This adds a "DefaultMixedHandler" class which will have the
highest priority (opposite the default handler). Initially,
it only supports the enter-pip+exit-split transition.

In order to support this, individual-feature handlers (split/pip)
needed to be refactored a little so that their animation code
could be re-used in a modular fashion by the mixed-handler.

Bug: 223321458
Test: atest ShellTransitionTests. Also, manual enter pip from
      app in split-screen with shell-transitions
Change-Id: Icf9152654d30028e62b0863cbdd2e917dde19c77

Former-commit-id: d5d644225638e916d822c328c93ff7cd09523ef8