Bubble flicker test: Ensure the notification shade is opened
This test depends on finding the text in the notification. In some
cases the notification might not be visible due to whatever else is
going on on the device.

This change adds a swipe that (based on local testing...) will trigger
the shade to expand while on lockscreen and allow all notification
text to be read / found by UiAutomator.

Update the text in the notification to use "BubbleChat" everywhere
so that it can be reliably found (when the notification is collapsed
or expanded it shows text a little differently & it auto expands
when the shade is expanded and it's the top notification).

Test: atest LaunchBubbleFromLockScreen --iterations 20
    - Ideally get the "app updates are ready" notif by checking for
      updates in settings, when you have this on screen the test
      would fail 100% consistently.
Bug: 205288792
Change-Id: Ib01923f35c531fb15230d640237e2c395ee1904d

Former-commit-id: d1f18027e715372c65cbe02eed915e66123ed4b2