Remove the unnecessary warning log
While cold-start an activity, a root task was created and
the activity type was set before adding the root task to TDA.
Therefore, the configuration needs to be updated and checked
for multi-window mode support. So, the warning log was dumped
since the root task was not added to TDA yet.

Bug: 163976519
Test: cold start an activity
Change-Id: I12f9167263939d760dabdd1098404fa9fa67e2c2

Former-commit-id: 3cb037b67e25a6afd43ff076bbee1d92a9cf2842