Send Split Info Callback before Activity#onCreate
The onTaskFragmentInfo callback from the server
triggered by the creation of a new activity may not
reach the client before Activity#onCreate is called
because of undeterministic IPC call order. However, the
application should get a split info update that includes
the newly launched activity before the Activity#onCreate
is called for that activity so that the application
knows that the activity will be in a split, which may
influence UI decisions.

Bug: b/218791665 b/200303250
Test: atest ActivityEmbeddingLaunchTests
Change-Id: If69c077db8ca390d41446e1329a448b3dd074293

Former-commit-id: d8659858914ba6d66885b4a8202f4876a054463e