Allow to update restart period dynamically.
We use a flag to control the restart period time, the value is static
and set in class level that means if we update the device config, we
should reboot the device to get the updated value.

This change reads the value in the constructor, so the value will be
updated when the voice interaction app updates no need to reboot.
Ideally it's good to listen the value change dynamically but it
increases the complexity here, we provide a workable and safe change
now.

Bug: 224618257
Test: update the package and the value is expected by command:
adb shell dumpsys voiceinteraction | grep RESTART_PERIOD_SECONDS

Change-Id: Id7d5f2d4f17de1f36baf6509598a82d6d73c5b53

Former-commit-id: b6e44763bd888f12142a3ac797848a36ef31e339