Fix seeing a weird window animation above TaskFragment transition
As DC#assignRelativeLayerForImeTargetChild will set relative layer for
non-IME focusable window to relatively above IME layer.

This layering setting could have a UI issue that during a
side-by-side TaskFragment changing transition, will see a weird
animation that non-IME focusable window is animating with being shifted
position on top of the TaskFragment.

In case seeing the window is animating above the app transition layer
because its relative layer is above the IME container on the display
area but actually not necessary, Add a check to not set the window
to be relatively above IME if the IME is not visible.

Fix: 225363555
Test: atest WindowStateTests ZOrderingTests
Test: manual as issue steps by using Google chat app with enabling
window extensions and side car library configuration on the device,
and verify if any weird window animation during launching in-call
activity fragment.

Change-Id: I641f54e18a18997f86779a355a23b17421fb35ec

Former-commit-id: 2ebbf7937435b2915304ca77d814b70ee39e40e3