Add new API to disable sending touch to wallpaper
Currently, all touch events on a Window are sent to wallpaper,
if the window uses FLAG_SHOW_WALLPAPER to show system wallpaper
in the background. This may have a security implications if
the Window is showing sensitive information. So, we are adding a
new API to allow developers to block touched to wallpaper.

Test: atest android.server.wm.WallpaperWindowInputTests

Bug: 177054831
Change-Id: I38d2e87d9b3b397f6946a638a7ecb4030ad21ddf

Former-commit-id: e9d766f333978a55bb128096063857aa8f2dfaea