Enable StrictMode checks for explicit garbageCollection from T.
This disallows calls to runtime.gc() in the new version of Android.
Explicit GC calls are blocking and can thus cause runtime overhead.
They used to be performed as a pattern in order to avoid GC_FOR_ALLOC.
However, that was only the case for Dalvik, that is now discouraged for
ART.

Bug: 3400644
Test: m
Test: atest StrictModeTest
Change-Id: I4720f2b4c4047bb90a88b924c539690e3c202e08

Former-commit-id: 8684d5be446100997524b198df6bc5994fd0f101