Make sure PiP transition triggered by WM is setReady
This was kinda working by luck previously. However, WM was
making a PiP transition in a way that it would never be ready
without other systems "accidentally" making it ready. This
results in time-outs. While this didn't pose any immediate
issues, its a potential source of some of the CTS timeout bugs.

Resolving this required adding a "push/pop NotReady" counter
to Transition so that we can have "code regions" which make
WM changes interleaved with performSurfacePlacement in one
transition in situations where we might need to nest into
an already-collecting transition.

Bug: 223321458
Test: atest RootWindowContainerTests TransitionTests
Change-Id: I5bc9c63b7f67b050e7164a6581b91eb2c0b00704

Former-commit-id: 4febc8a8ffa76bda6ab9ab33a56700d8a0b4cab8