Do not apply taskbar insets for overlay case.
Taskbar is shown as an overlay when an app is in immersive mode. The
user can swipe away the taskbar, or swipe up to show it. No insets need
to be applied in this scenario; the app content can stretch to the
bottom of the screen.

Also, migrate from InsetsState#getSource to InsetsState#peekSource to avoid unnecessary creation of the source.

Fix: 202836652
Test: manual
Change-Id: If573b2fa3936b81f92a2c196d49149782cb65b0a

Former-commit-id: c6b845dee35adff5498fe742b0a845cdf4f20b2f