Add displayId to transferTouch
To account for the case where the displays are mirrored, add display id
to transferTouch(..).

When mirroring is on, windows with identical tokens would be positioned
on different displays. This could break the transferTouch API, if the
window of mirror happens to be found first. This could mean that
screenrecording may have different behaviour from the regular usage.

Bug: 220077253
Test: atest inputflinger_tests
Test: manual testing - pull down notification shade from home screen
Change-Id: Ia8965cad83e2c29062329958edbc9516ae67f1ed

Former-commit-id: 57142133670c0d6fd92548ecb230bb0ad6547974