Show "review notification permissions" notification
This notification is shown for users who have XML version <4 (newest for now), which includes both users newly upgrading from <T and anyone who has already upgraded to T but not yet seen the notification.

This state is tracked using a Settings int that allows for coordination between PreferencesHelper (which reads XML), NMS (shows notification), and the new ReviewNotificationPermissionsReceiver (receives intents from the notification and behaves accordingly).

Change also includes a new JobService in order to schedule the job to re-show the notification 7 days in the future if the user requests it.

Fixes: 225373531
Test: manual; NotificationManagerServiceTest; PreferencesHelperTest; ReviewNotificationPermissionsReceiverTest; ReviewNotificationPermissionsJobServiceTest

Change-Id: I4039728646f706bdc1c01f2c0165721a5b10c9f5

Former-commit-id: 5dfa2dbca6e810102102bbdff7d6210d8f0af3f3