Do not allow panel expansion with bouncer over dream.
This change prevents the PanelViewController, the parent class of
NotificationPanelView, from expanding on end motion when the bouncer is
shown over dream. This prevents the panel from appearing erroneously on
motion cancel during swipe down on the bouncer. This cancel comes as a
result of the bouncer pilfering the touch events once the gesture has
been detected.

Fixed: 228347831
Test: atest BouncerSwipeTouchHandlerTest#testInformBouncerShowingOnExpand
Test: atest BouncerSwipeTouchHandlerTest#testInformBouncerShowingOnCollapse
Change-Id: Ia668954d31e26d704f9827b378a7ab0724a6ec4f

Former-commit-id: 5cb83b89ddc9d581d392b76b347ed53d09e28e66