Improve Condition Monitor threading.
This change modifies condition monitor to enforce all modifications to
the callback listen happen on the same thread. First, the executor now
is specified to be the @Main variant. Secondly, callbacks can no longer
be added in the constructor. Only allowing callbacks to be introduced
post construction enforces this modification is always on the main
thread.

Test: atest ConditionMonitorTest
Fixed: 228135569
Change-Id: I376479d104e0252eb3a39d768299ce86a78e11a1

Former-commit-id: cb10414f11e8488f757fb7040d83bff281543502