Reset embedded TaskFragment override when entering PIP
When an embedded activity is entering PIP when it is the only activity
in the Task, the whole Task will enter PIP.

When this happens, we want to reset all organizer override config so
that the activity will get the correct config. We don't want to remove
the TaskFragment because we can use it to get the windowing mode change
in organizer.

Besides that, we will disallow any further organizer override to the
TaskFragment in PIP.

Bug: 225371112
Test: atest WmTests:RootWindowContainerTests
Test: atest WmTests:TaskFragmentTest
Test: atest WmTests:TaskFragmentOrganizerControllerTest
Change-Id: I03bd679b77f7ed8a65ba5fc59723b83a4254442c

Former-commit-id: bc357abca6ef0e0ebe4ec18b3986d1d4351ba0a7