Guard against incorrect ID assignment in CdmService
Do not assume that all previously assigned association IDs were assigned
correctly in regard to ranges of associations ID allocated to each user
when generating an association ID for a new association in
CompanionDeviceManagerService.getNewAssociationIdForPackage().

Also Slog.e() when detect such errors in existing Associations as well
as when detect ID conflicts in AssociationStoreImpl.addAssociation().

Bug: 224736262
Bug: 223995714
Test: atest CtsCompanionDeviceManagerCoreTestCases
Change-Id: Ibe14cfda9718b3e473f0f157bbfef2f7051819c2

Former-commit-id: 8228658b691897f110c5d77195a97b2f2f2b2aeb