Add rate limiting to addTombstoneToDropbox.
Support for attaching a
header for the number of dropped entries for tombstones in proto format
will come in a later change.

Use the .pb tombstone file for both plaintext and proto tombstones as they both should always exist and the proto one is easier to parse. The plaintext file needs to be kept around for gmscore needs.

Bug: 225173288
Test: atest ErrorsTest

Change-Id: If6434e4c309d010624a15d7b478bc645c326e8ef

Former-commit-id: 7c47347a17a153ba1593fbd2157b9f93ae4625a5