OMS: fix broken 'overlay list' output
Immutable overlays that target the android package are incorrectly
omitted from the output of 'overlay list'. However, those overlays are
correctly omitted when the package manager is told what overlays are
enabled.

Move the filter closer to the interaction between the overlay and
package managers so as not to interfere with other operations such as
'overlay list'.

Bug: 228404475
Test: atest OverlayHostTests OverlayDeviceTests
Test: manual: 'adb abb overlay list android' includes immutable overlays
Test: manual: inspect /proc/<pid>/maps to verify idmap files are not loaded twice
Change-Id: I94f5206326fef6c3b62d6d1c7a9e74ea161495ee

Former-commit-id: 5dba5f93d445abf6c559e08e5c2c60c13feffb50