Fix bubble position with IME when the bubble has a custom height
Usually bubbles are full height, but it is possible for a bubble
to have a smaller height. In this case the pointer didn't align
correctly when the IME came up.

This change makes it so that the expanded view will translate up
a bit with the bubble (only when the view isn't full height and has
space to translate), which keeps it aligned with the pointer.
This is a bit of a nicer experience because the IME doesn't overlap
as much of the app as it would before.

Test: manual - have a bubble that has a small height, focus the IME
             => verify the pointer follows the bubble
Bug: 223906486
Change-Id: I1542f06887ccc8370a762fc6ebe42aca26559cb4

Former-commit-id: d303a0262785c2d1da3bb7ab354cdbddcd19f5d4