Modify handling of notifs with invalid sounds
Previously, if an app posted a notification with an invalid sound
and they targeted P+, we would not post the notification.

However, sounds are often not in the control of apps, and apps
should not be blamed if those sounds get corrupted.

With this cl, rather than dropping the notification if the app
provides an invalid uri, we will replace the sound with the default
notification sound.

Bug: 227770797
Fixes: 201488127
Test: NotificationRecordTest; modify sms sounds to be invalid and ensure
notification is posted with sound

Change-Id: Id438b7ea6912b13cfb79a09fe700474ad17879de

Former-commit-id: 2210ff10e368d6619971a2306ea93e752e169549