Update task manager ui information on dialog opening
A proper fix would be to have a listner for changes to
getBackgroundRestrictionExemptionReason(), but creating this listener
mechanism would be very difficult since the value is computed on demand
and there's a lot of signals that go into it.

This change works around by refreshing the ui information when the
dialog is opened. In theory we should also update the values when the
footer is shown but this will cause a lot of IPCs and with the current
policies the ui in the footer won't actually change as a result of the
exemption reason code changing.

Test: Change the battery restriction and validate UI has changed.
Fixes: 228306315
Change-Id: I61857df989ea6015c91dbd60db1d700a1dde266e

Former-commit-id: 6336ca4738bfb4e580ccde8f4bb6353cb2d3e666