Reland "Make extra navigation bar count towards non decor frame""
This reverts commit d3766eb4c94828773abb8398cb9a9af9d07e0e2f [formerly 8626ad35b186d20a67bdb30119c4837fc08bff25].

Reason for revert: Trying to reland

The previous patch may delay a configuration update caused by the task
bar change, and dispatch the update when something irrelevant happens
such as dragging an icon to the workspace. Changed logic in WMS and
WindowState to ensure a configuration update in that case.

Test: TaplTestsQuickstep#testAllAppsFromHome (With task bar)
Test: CtsWindowManagerDeviceTestCases (With task bar)
Test: android.view.cts.ViewTest
Bug: 224404595
Bug: 205496695
Bug: 228380863
Change-Id: Ib192d4ad012fed136f808b15a84f8d83ca426b21

Former-commit-id: 620e91578cb5d0bbfbf6f8c5aedb77e23c13ef49