SystemUI: remove unused code that showed radio status in CryptKeeper mode
Since support for Full Disk Encryption was removed in favor of
File Based Encryption, "CryptKeeper mode" no longer exists and the
"vold.decrypt" system property is never set.  Therefore, remove the
unused SystemUI code that showed the radio status in CryptKeeper mode.

This is, very roughly, a revert of ag/1350315 which added this feature
in 2016. However, the code has changed a bit since then.

As discussed with Eric, I'm cherry picking this CL in master to tm-qpr-dev to mitigate a merge conflicts in a CL of mine and also to prevent the two branches from diverging.

Bug: 208476087
Test: mp droid
Change-Id: I6231762e8353113366eba92b686d6d61582fb2b7
Merged-In: I6231762e8353113366eba92b686d6d61582fb2b7

Former-commit-id: 999f98feb72ce1ffda94a964be7fce0dc4a6d746