Allow null prev/next buttons to be INVISIBLE instead of GONE
Added an overload method for setVisibleAndAlpha that allows callers to
specify the value for null as INVISIBLE instead of GONE and called that
instead for the prev/next buttons. Added booleans to MediaButton to
allow it to say whether to reserve the space for the prev/next buttons
when they are null. Made the MediaButton class immutable.

Fixes: 224749799
Test: tested manually and added 2 tests to MediaControlPanelTest
Change-Id: I5ccac017eccfb2dc2ecdd42a1e6517d4787c68b6

Former-commit-id: 5304fa7896597de7a78803123decdfdfe81bfd93