Watch uid proc state instead of importance for 1-time permissions
The system process may bind to an app with the flag
BIND_FOREGROUND_SERVICE, this will put the client in the foreground
service importance level without the normal requirement that foreground
services must show a notification. Looking at proc states instead allows
us to differentiate between these two levels of foreground service and
revoke the client when not in use.

This change makes the parameters `importanceToResetTimer` and
`importanceToKeepSessionAlive` in PermissionManager#startOneTimePermissionSession
obsolete.

Test: atest CtsPermissionTestCases + manual testing with mic/cam/loc
Bug: 217981062

Change-Id: I7a725647c001062d1a76a82b680a02e3e2edcb03

Former-commit-id: 618641451d724572597024cd5ced7e46fc801c2c