Assert correct nav bar visibility in overview
The bars (including status bar and navigation bar) are expected to be hidden while entering overview in landscape if launcher is set to portrait only.
Because "showing portrait overview (launcher) in landscape display" is an intermediate state depending on the touch-up to decide the intention of gesture, the display may keep in landscape if return to app, or change to portrait if the gesture is to swipe-to-home.
So instead of showing landscape bars with portrait launcher at the same time (especially return-to-home that launcher workspace becomes visible), hide the bars until leave overview to have cleaner appearance.

Bug: 227189877
Test: atest FlickerTests:OpenImeWindowToOverViewTest
Change-Id: Idc89eab44bb74217c4235374a8fcfd5f7ec488cf

Former-commit-id: c63675f595383b78a6b97d4b5a5d2f57506c85cb