Finish unfold transition when screen turns on after half-folded
When screen turns on after we received half-folded event
the transition started and didn't end. We had the same issue
when the fully unfolded event was received before screen
turned on. In this case now we forcefully finish the transition.
I added the same behaviour for half-folded case as well.

Bug: 222705041
Test: quick folding/unfolding with only
 0/10 degrees hinge angle sensor
Test: atest com.android.systemui.unfold.updates.DeviceFoldStateProviderTest
Change-Id: I964217acaaba77926c43796d16ad00a4ca15203e

Former-commit-id: fd355c0945a0524307c9d3d2e30fe7ae4330e029