Resend remaining VMAs on partial compaction due to failure
process_madvise syscall returns the total bytes that were processed
however, upon failure it will send the partial amount of bytes
that were processed in which case this patch aims to skip the failure
page then resend any batch of vmas that was not fully processed.
This way we would effectively process the whole range of vmas
rather than skipping the rest of the vmas in batch when
process_madvise does a partial advise.

Test: Manual
Test: am compact <pid> <uid> full
Bug: 205658049

Signed-off-by: Edgar Arriaga <edgararriaga@google.com>
Change-Id: I6445ea85f5eb028cb1d11a4af77c5c4cedf76954

Former-commit-id: 33b1957c88dcbf1efb7a5140673dbf0d83dbffea