[User Switcher] Finish activity when switch user
When switching to owner (user 0), we don't end up in lock screen like
when you switch to a secondary user. My belief is that this is because
the system user does not need to spin up a new process for system UI,
whereas a secondary user will need a secondary process.

We can keep the experience consistent by ensuring that the activity
finishes when a user is switched.

Bug: 228901638
Test: Manual Test
Change-Id: I74623655704cb89e499940498aef5a811dadf77e

Former-commit-id: 848a7b8d357b1af3a5e8f64a7cb1a214644952ce