Show Notification Permission dialog on (some) task trampolines
If an app creates its first notification channel, and the current task
was trampolined from another by the same app, then show the notification
prompt. If an app starts, and is a trampoline, check if there are any
other running, focused tasks which were a launcher start. If so, show
the notification prompt

Test: manual
Fixes: 229282386
Change-Id: I2c2b96b1597d7f09a44668423308ae76d1f9699c

Former-commit-id: b17596b05127cae7a15d6d127c42bb4501b2c731