V1 is the latest version of android.media.audio.common.types
As part of the AIDL API freeze for TM, the V1 of
android.media.autio.common.types was frozen. As a result,
unversioned reference to the interface now refers to the V2
of the interface, yet to be frozen. This causes ODR violation
where V1 and V2 of the interfaces are both used in the
platform.

To fix that, replaced the unversioned reference explicitly
with the reference to V1.

Bug: 225941299
Test: m
Change-Id: I1d4faa3cd541acdecacec5c711050d0e925ca36d

Former-commit-id: a031a2940a6e872c2ae96996d723821fa907adf6