Construct Region object with nativeObject
The default ctor for the Java Region object creates a native SkRegion
object. Hoewver, in the JNI code we overwrite the native pointer so the
native object created from the default ctor never gets cleaned up.
Instead, use the ctor that accepts a native ptr so we create the Region
object with native info populated.

Test: Builds
Fixes: 229537097
Change-Id: I89e4b8ec985c538e7337c0e80415d3b6474da78e

Former-commit-id: db07b07b2b198f0bdc7bfb817b35410970b65f3c