Remove NOT_RESTRICTED from basic network requests.
JobScheduler only ever tells a job to run on the default network. If the
connectivity stack says an app's default network is a restricted
network, then the app has permission to use that restricted network.
However, default NetworkRequests exclude restricted networks by default.
Removing this exclusion for basic job NetworkRequests so that a job can
run on a restricted network if the app is allowed to do so.

Bug: 219788183
Test: atest CtsJobSchedulerTestCases:JobInfoTest
Change-Id: I3c2b9f725c7e1254c7708f5b8df7bdf18163608e

Former-commit-id: d8ff97bf8d33d71a098733b0e241290c2fbe53eb