[DO NOT MERGE] Force extra nav bar provide side gesture insets
This is a patch to fix the missing gesture insets issue when there's a
task bar and no navigation bar on the display. It will try to add the
gesture navigation insets when a window providing navigation bar insets
is added and there's no window with TYPE_NAVIGATION_BAR.

The logic will only be applied when the insets is not flexible. Due to
the removal of the old hard-coded logic in the later version, this patch
should not be merged into the next version.

The final fix will not be on tm-dev to keep tm-dev stable. Cherry-pick the hard-coded solution to avoid failures related to IME.

Bug: 231959399
Bug: 229805460
Bug: 232044523
Test: android.widget.cts.SeekBarTest#testSetOnSeekBarChangeListener
Change-Id: I080df3547361aef1aa259dd2b3dc9bdb860a0193
(cherry picked from commit 798a3ce783e1166597942361717c42cdd5a0cff7 [formerly 1921f8c1dc660851a499c5dac5a55cedf854f93e])

Former-commit-id: a2e8e229719aa0d3ff0aa8bd50b56fd569628ca5