Fix keyboard is attached to wrong task in split
We introduced CL[1] to fix but it may cause a timing
case that control target has already updated but the
reparent in DC#updateImeParent is ignored because
ImeLayeringTarget and ImeInputTarget are different.

Also next updateImeParent won't be executed since the
Ime control target didn't change.

Attempt to fix this by force update the Ime parent if
the control target is remote and the IME parent is not
yet going to update to the remote one.

[1]: I332c0e4fff62df5d7b793eda2767bb58fe85a938

Bug: 228791170
Test: Manual test
Test: atest DisplayContentTests#testComputeImeParent_remoteControlTarget
Change-Id: I834568231a8affbce55dc44c17fa0086f7999e35

Former-commit-id: e27e307baaf033749ae301328379ea6cf6d46c59