[MultiUser] fix async bouncer user
The userswitcher in the bouncer does not update when receiver in the
controller receives the user switcher intent. This change will add a
callback to the aforementioned controller and we can reload the user
switcher upon doing so. With some testing on the device, there does not
seem to be any flickering, but bug is resolved.

Previous CL that was supposed to fix this b/17716871

Bug: 228901381
Test: Manual on device
Change-Id: I7b9fada327db63d744c340f75398c719b20ef392

Former-commit-id: 1e5c6d9c4e7991fde18993bfccabb4207d8977e8