Do not change created-by-organizer task to different organizer
Otherwise when registering another organizer, all tasks will belong
to the latest organizer. And since commit 4d63b2b, it will remove
all created-by-organizer tasks when disposing, which also removes
the tasks created by other organizer. That causes the original
organizer to have unexpected state or crash when it becomes the
active organizer.

Remove parameter forceUpdate because it is no-op (setTaskOrganizer
also skips for the same instance). It was used in commit 77338ab.

Bug: 228977409
Test: atest WindowOrganizerTests# \
      testUnregisterOrganizer_removesTasksCreatedByIt
Test: atest AivityLifecycleLegacySplitScreenTests
      And enter split screen from recents.
      SystemUI should not crash.
Change-Id: I1c495717c536f021a98b338189b7924630dcc7f3

Former-commit-id: eebc3162a2d7d8dad225cc7c52e5bc3ab200037a