apk-in-apex-cache 1/n: Fix isCacheUpToDate for apk-in-apex
In case an APK resides on the APEX mount point it's mtime will always be
zero. This breaks the isCacheUpToDate check, which is based on comparing
the mtime of the apk to the mtime of the cache.

This cl fixes the issue by using the mtime of the APEX file backing the
/apex mount point the apk resides on.

Also noticed some inefficiencies in ApexManager and added TODO's to fix
them in the follow-up cls.
s
Bug: 225435110
Test: atest PackageParserTest
Test: atest ApexManagerTest

Change-Id: I8a0ab8952454652048076a26e81097b95991ae60
Merged-In: I8a0ab8952454652048076a26e81097b95991ae60

Former-commit-id: 2589a7e8f756ce3212b6763dfc5cb91fe04e50ec