Fix notification shelf issues
NotificationShelf/StackScrollAlgorithm:
Define "shelfStart" as yTranslation of space before the shelf
because that is the point where we no longer see the notification
being clipped by the shelf.

StackSizeCalculator:
Add missing divider before first element.
Currently, all elements have divider before them. This is not ideal
as we should be saving vertical space when possible - this will be
addressed in a future change since this change is focused on fixing
the mismatch between StackSizeCalculator and StackScrollAlgorithm.

NotificationIconContainer:
Remove overflowWidth since it is always a single iconWidth.
We can use iconWidth instead of keeping this unused scaffolding
for showing multiple dots.

Fixes: 227361193
Test: NotificationStackSizeCalculatorTest
Test: NotificationIconContainerTest
Test: on lockscreen, have 4+ notifications, swipe them away
      => shelf updates icons as expected
      => shelf does not only show a single overflow dot
Change-Id: I431a00c78a26186446182ef91f7b8935f5f422d2

Former-commit-id: 50f61bfdb74c44c258f35194c3567748a34ba1ad