Update cache key calculation for statusbar insets
We use Display#uniqueId as a cache key for status
bar insets when using multiple displays (e.g. foldable
device). But uniqueId and display size could be
inconsistent (see b/198965093#comment13): we could
receive uniqueId for a different display than the
current display size in the configuration.
So if it is calculated once with wrong display size
it is saved in the cache and status bar is rendered
with wrong insets every time.

I updated the cache key calculation with the actual
inputs of the calculation algorithm (display size
and cutouts) which are consistent when
onConfigurationChange received.

Bug: 212924195
Test: folding and unlocking a foldable phone >1000 times,
 rotating when folded/unfolded
 => checking that the statusbar has correct size
Change-Id: Ibd8575529b031747a802b56db3eb12069b9ba6b0

Former-commit-id: c0e76e7cd8998354347b838822ecbe4c7b9ddf1f