Deprecate and no-op test API getContentCaptureServicePackageName
After the fix for b/222231605, permission pre-grants are no longer tied
to the package that provides ContentCaptureService, so this test API is
no longer needed. However the API is still referenced from GTS tests
(though it won't be invoked on T and above), so we need to keep the
method in the SDK to compile the test code.

Bug: 229662628
Test: Presubmit
Change-Id: Iddc201ba8afde816446377e509a3f2de35215f61

Former-commit-id: 22f3baa869a134b9a9ffffcb77a8d7c5bc8a0fb1