Optimize ActiveServices targetSdkVersion fetch
Rather than fetching the entire AppInfo object, call the method for
specifically getting the targetSdkVersion.

In this specific case, match flags and user ID can both be ignored as
only 1 physical APK for each package name is allowed on the device, so
as long as the caller has package visibility, the targetSdkVersion
returned is guaranteed to be the right one.

Bug: 219481486

Test: presubmit

Change-Id: I1d8c155e9be9b80da0af241273ceb9cf3564d917

Former-commit-id: 9aedadbedc955e854f052e69ba07a4e6a0ad688f