Advise use of InputDeviceListener when using pointer capture
InputDevices can be reconfigured when pointer capture changes, but the
app will be notified via InputManager when the updated device
information is available. This races with the onPointerCaptureChange
callback, so advise the use of InputDeviceListener in the pointer
capture documentation for developers to get around this.

Bug: 226425883
Test: None
Change-Id: I7aeb23722dc40b8fd37cb9667a4f23a336e8b5f1
Merged-In: I7aeb23722dc40b8fd37cb9667a4f23a336e8b5f1

Former-commit-id: 98e54bb941cb6feb07127b75da37833281951d52