Schedule lifecycle-stops after every transition
Was trying to limit it only to transitions with "closing"
activities; however, this doesn't work if a new transition is
created before the relevant transition finishes (eg. during
fixed-rotation). So, just always attempt to schedule. This
should be ok because the schedule method already no-ops if
there are no pending-stopped activities.

Bug: 230531683
Test: atest AppConfigurationTests#testAppOrientationRequestConfigChanges
Change-Id: I82cc9dc51291e727736d2b6a6feda1a6d171d235

Former-commit-id: 0683868c00d060a45427c01dc0fe3cbcef73594e