Respect allowAllRotations for rotation suggestion button
AllowAllRotation config flag controls if it is allowed
to perform rotation in all directions (including
upside down). This config is true for tablets to allow
upside down landscape rotation. When auto rotate is on
this allows to perform this rotation, but when auto
rotate is off, rotation suggestion button
didn't respect this configuration.

Adding allowing to show the rotation suggestion button
when this configuration is 'true'.

Bug: 229565988
Test: atest WmTests:DisplayRotationTests
Change-Id: I62fd2f49b87bf1739baac968581f94f9b9b181a8

Former-commit-id: 624b03d6cd72e17d2ecd5533a0c60fce6b227562