Only run Animation#setBackdropColor on overriden app transition
Only setBackdropColor from overridePendingAppTransition when mNextAppTransitionType is NEXT_TRANSIT_TYPE_CUSTOM since that is when we want overridePendingAppTransition to take effect. And because mNextAppTransitionBackgroundColor is not cleared after overridePendingAppTransition is called be don't want to continue applying it to other transitions leading to a persistent background color. And in some cases a is null which leads to a NPE if overridePendingAppTransition was called at some point before.

Bug: 229929107
Bug: 229822876
Bug: 222438428

Test: atest CtsWindowManagerDeviceTestCases:ActivityTransitionTests
Change-Id: I39c768f147a0fafec063c3929974c280556997cc

Former-commit-id: 96b645d043dfc296d2c83db1db4ffa74d500e244