Choose the type of callback based on the latest callback
Connectivity status change may cause icon changes in the QuickSettings
tile, but there is no specific callback for it; instead the tile will
reuse the callback for Cellular or WiFi status change. We should choose
the type of the callback based on the latest callback type to avoid icon
flicking.

Bug: 220069227
Test: Manual tests, Unit tests
Change-Id: I01afb71b57a54f94cadf50bcea64b6da2b3ec833

Former-commit-id: a41dcf56d99e99f0f42f545235fe3a7cad8abded