Scribe: Find candidate view in stylus down event
In the initial implementation, HandwritingInitiator looks for a
handwriting candidate view when a stylus motion has exceeded the touch
slop distance.

This updates the implementation to look for the handwriting candidate
view in the stylus down event. The overall resulting behaviour is
unchanged.

This is needed for an upcoming change to stop dispatching MotionEvents
to the view tree if they are for a stylus movement on a handwriting
candidate view, in order to prevent scrolling before handwriting is
initiated.

Bug: 228210945
Test: atest CtsInputMethodTestCases:android.view.inputmethod.cts.StylusHandwritingTest
Change-Id: I395c2ec4b16a3fd3cb56308da872ec959e93943e

Former-commit-id: 08a94de0dea5e7fb94042acad29f51ddd05bb9d3