Update when PlaybackState state or actions change
Some apps are updating PlaybackState without also posting a new
notification, which is looks buggy when using the new actions.

This adds a check so that when the PlaybackState updates, if the actions
(standard or custom) have changed, or the state has changed between
playing or not, MediaDataManager will get notified
and update the PlaybackState-based buttons.

Fixes: 214335602
Fixes: 229392609
Test: manual - see bug
Test: atest MediaTimeoutListenerTest MediaDataManagerTest
Change-Id: If2eb02c9c969124ebf193a89ff37edd5b1e6ede1

Former-commit-id: ecf1d11f7234886d2affb3764cf8fbd0e110bd6f