Fix wakeAndUnlock immediately after sleep.
This resulted in fragments of the shade/keyguard
being visible as we unlocked, due to the screen
off animation putting the shade in a half
KEYGUARD half SHADE state. The existing code
assumed that wake and unlock would always be
from a fully settled KEYGUARD state.

By simply updating the keyguard state as soon
as we wakeAndUnlock, things look good.

I tried to add a test in CentralSurfacesTest,
but StatusBarStateController, etc. are not
initialized because that test is not set up
correctly.

Fixes: 229047510
Test: atest SystemUITests
Change-Id: I0603d2c1d467c8172764ff23be1c3879b8f2edfe

Former-commit-id: aa96663319a05008731d39a0d367144fbe38599f