Reland "Allow activity in ActivityEmbedding to get Task windowing mode"
This reverts commit 9207ac1173518a0dffaf65d8b1d8236e422ff249 [formerly 0dc147eeeef9a8583a93b6f0933294c0ec502ff2].

Before, when we are in ActivityEmbedding split, the TaskFragment
windowing mode is always set the WINDOWING_MODE_MULTI_WINDOW, which
overrides the activity windowing mode, and it causes the app process
unaware of the Task windowing mode. Now, when the Task is multi
windowing mode, we set the TaskFragment windowing mode to the same as
the Task.

Bug: 230446204
Change-Id: I61b0857c551e6fd4ee9a544b63a3a2c20e291fab
Test: atest WMJetpackUnitTests:JetpackTaskFragmentOrganizerTest
Test: atest WMJetpackUnitTests:TaskContainerTest

Former-commit-id: 10e4a19ef4be6471e826c951e2136e7bc3eab61a