SpatializerHelper: fix init of compatibility for attached devices
 SpatializerHelper.init() is called when AudioService starts, but
also after each audioserver restart. After evaluating the
transaural/binaural capabilities of the spatializer effect,
the associated devices were added to the list of compatible
devices regardless of whether they were already present, this
was causing duplicated entries in the list.
 Also avoid always setting the enabled field in the list of
compatible devices to true if the device was already in the list.

Bug: 230871496
Test: repro in bug

Change-Id: If76658677b44d22414fa393208b69d8ec82592fd

Former-commit-id: 2ec2ec5a6bc247ededb6a22904a1b761c29a73ed