Add a new binding flag BIND_TREAT_LIKE_VISIBLE_FOREGROUND_SERVICE
Repurpose the legacy BIND_VISIBLE to this new flag, as the BIND_VISIBLE
was actually never implemented.

This new flag is for use by the system or persistent processes,
the process state of the service process will be bumped to
the PROCESS_STATE_FOREGROUND_SERVICE, with oom adj score "visible".

Bug: 230897753
Test: atest MockingOomAdjusterTests
Change-Id: If86ec9281ec482a0916c78338ae88cf2199256e4

Former-commit-id: 84f00cf8afa4658997a8bf128623b205ef9c8312