Add traces of the size of HWUI shader cache in RAM
When HWUI needs to either load a shader from or store one to persistent
storage, this means a new shader is in its temporary (RAM) cache. Keep
track of the number and include it in perfetto traces.

As long as we're modifying ::store, make it override the version with
the newer signature, since the old one is deprecated.

Bug: 231194869
Bug: 225211273
Test: manual

Change-Id: I36b7e018291c50ed315f534f15946047ca001300

Former-commit-id: 8cedb66717694afe6584edefc0ed95c5dd18f8ed