Don't request sync again for reparented WC
When a WC is reparented, a sync request can be initiated again to
account for any config changes. However, this can cause deadlocks in the
client since the app may already have produced a sync buffer and that
buffer is now held in a sync transaction. The app can't produce any more
buffers since it has ran out and just gets stuck waiting. The previous
buffer will never get applied since BLASTSyncEngine is waiting for the
next buffer to return from the sync request.

Disable the second sync request for now and handle a real solution
later.

Test: No ANR in split gmail + messages on 60hz device
Bug: 214320347
Change-Id: I5e11925de848781ddc34793f118b739703cc8ca2

Former-commit-id: 63238fe68480ca6a3ecd31e6e3edb1c24c949e21