Add new methods to activate/deactivate session
Currently the way that identifies an active session of ApplicationsState
is checking a variable. However the variable is binding with a task that
loads all apps. There is no appropriate way if a class that implements
Callbacks just needs to receive the callback without loading all apps.
The current methods might trigger the task that loads all apps several
times if we would like to activate a session. This CL is to add seperate
ways that enable/disable a class that implements Callbacks to receive
the callback call.

Bug: 222985623
Test: manual test
Change-Id: I46c7143ddf220cbe7c50ee124aabe63f75be03ee

Former-commit-id: 2379385a09f5ef590681c51caa7ab4cced01cf87