Improve LogAccessDialogActivity's rotation handling
1. Do not dismiss the dialog on configuration change, such as rotation.

When the device was rotated, the dialog was automatically dismissed as
the activity was destroyed & recreated in the new configuration.

2. Wrap the dialog content view in a ScrolLView

In landscape orientation (and smaller phone screens), the dialog content
may not fit entirely on the screen. By wrapping it in a ScrollView, all
the dialog text and buttons can still be accessed in landscape / smaller
screens.

Bug: 231436416
Test: Manual, on adt3
Test: On phone emulator in light/dark mode & different orientations
Change-Id: Ibf3ecf93a237783ce4909a24a959c3ea243ab766

Former-commit-id: bd9cf8f9d2df3dea41513fa2061c5f8920d44a5b