Allow profiles of an admin user to remove system app updates.
This modifies a change made earlier that prevented non-admin users
from removing updates to system apps. This was intended to prevent
users such as guests from reverting updates, but an unintentional
effect was that profiles of the admin user (such as a work profile)
also became unable to do this.

This change allows a profile to revert updates if the parent is an
admin user.

Bug: 229719733
Test: atest DeletePackageHelperTest
Change-Id: Ie9065dfa6a0117f8cf1f49382a8e44655b17cec1

Former-commit-id: 5bd7689ddddf74b5f4f97bcab4e3c1f2a5011d06