[Media] Update player's a11y content description based on whether guts is open or not.
Note: I was hoping that just setting the content description to null if
guts was open would work. When I tried that, it *did* announce "Hide this
control" if I tapped on the player via TalkBack, but it *did not* announce
anything if I used navigation swiping until I was selecting the media
player. So, this code just always sets the description.

Bug: 229574744
Test: manual
Test: MediaControlPanelTest
Change-Id: Ifa69521289a4b7a5cd7a35f092a76827a560e3b1

Former-commit-id: b5c0f6c60ca63b1f5b0be2940b2d8a21d0116185