[Partial Screenshare] introduce a hidden, permission-protected Activity API
Hidden Activity API allows a component holding permission
MANAGE_MEDIA_PROJECTION to declare that the result of setting up the
MediaProjection session should be returned immediately to the host VC app.

The host VC app will cycle through the resumed state when the activity
result is delivered to it, and return to its original state.

Manually tested with test MediaProjection app.

Test: atest WmTests:ActivityRecordTests
Bug: 230607871
Change-Id: Ic85fe8657b9e7f526bbc37a52011d16afa8aef2c

Former-commit-id: 12f5992e4df6b5948cf456306213e59c149f8c45