Destroy the last embedded activity immediately
When a back gesture on the last embedded activity, it is added to the
stopping list and waits for an idle event to be finished. But, the
next activity, on the other splitted side, is already visible and
won't trigger the idle event. Eventually cause the split-screen cannot
be dismissed.

The embedded activity doesn't need the animation because it is the last
one. Destroy it immediately so split-screen can be dismissed.

Bug: 230588790
Test: atest RecentsAnimationControllerTest
Change-Id: I4c230cfe902531db7d80b3bbe00e89c30ced99aa

Former-commit-id: 0c893536cf8453316b0ccab9468d0d1e1d6f2116