Change a nullable get to getOrDefault(,false)
Implicitly casting from a Map<,Boolean> to a boolean will cause a
NullPointerException if the map isn't populated yet. The default
should be false because this is more conservative with respect to
how users are allowed to unlock their device if the real information
isn't available yet (or won't be available because the device doesn't
have a keyguard yet still calls this API).

Test: adb reboot; adb logcat | grep NullPointerException
Fix: 231987867
Change-Id: Ie631286a30eaa1d9db36d8a149b9165df292429d

Former-commit-id: 93a98d862fbf90af9e3b5945d2ca613d8ec62fe9