Using isUidPrivileged instead of getPrivateFlagsForUid API
Starting from U, the PackageManager#getPrivateFlagsForUid does
not support to get private flags for different users. It
returns nothing if the package cannot be found in the calling
user.

To avoid failing to get private flags in the IntentFirewall,
this CL uses the internal API isUidPrivileged to check the
privilege state of the caller.

Bug: 229684723
Test: Build
Change-Id: I8440cb51e4318d57a097efe651be626b9df2da0f

Former-commit-id: 87944e742cea0890f686126fd8e706a11c6155d2