Fix package visibility broadcast regression
Sending broadcasts are not restricted by the package visibility of the
calling package. This is achieved by clearing the Binder calling
identity in several top-level broadcast sending APIs.

ag/15315839 introduced intent filter matching enforcement, which
requires the real calling UID to be used for evaluation. This caused
regression in broadcast package visibility.

Test: manual
Bug: 230363029
Change-Id: I6bf34fe22aa9396786e3bbbfe64094474c34b6df

Former-commit-id: 8b1f96741ccb49ec04ed2b4b1334153a4935834a