Don't send redundant WCT for exit-pip finishTransition
exit-pip reaches the end state at the beginning of the transition,
sending another WCT at finish can erroneously interleave with
other transitions (esp. other pip transitions) causing the system
to end up in a bad state because it can effectively re-apply
the expand-pip WCT after a transition it was originally ahead of.
It is also redundant.

Bug: 231150615
Test: atest PinnedStackTests#testPipUnPipOverHome
Change-Id: Ica657d7d610226a09f34375685ed4089a63d4e92

Former-commit-id: 371717ffcac2e5342d318141f4408ddc36b65596