feat(magnification): move mag button when window reaches the screen edge
Wait to move it until drag is done, if they are dragging by the handle.
But if it is moving via two-finger drag or triple-tap, update the drag
handle immediately in real-time.

Add tests to verify applyTapExcludeRegion() respects the drag handle
position. This required fixing a bug in earlier code where it was adding
extra padding around the drag handle. Now we only exclude the tap area
of the handle, no padding.

Bug: b/205904754

Test: unit and manual:
1. Drag the magnification window to the right edge of the screen. The button should swap to the left.
2. Tap the button. It should bounce the frame.
3. Drag back to the center. The button should return to the right.

Change-Id: I04af438ad62359ac31ba12aee0a1739e25068937

Former-commit-id: 205b1478cc602cddb148eaceab6d54428f08a8c3